move type, remove unused function
