+++ app/api/document/patch/route.ts
@@ -1,12 +1,6 @@
-import type { FileQuery } from '@/tina/__generated__/types';
-import type { Placeholders } from '@/types/index';
+import type { PatchBackendParcel } from '@/types/index';
 import { NextResponse } from 'next/server';
 import patchDocument from '../../../../utils/patchDocument';
-
-export type PatchBackendParcel = {
-  file: FileQuery;
-  placeholders: Placeholders;
-};
 
 export async function POST(request: Request) {
   const patchBackendParcel: PatchBackendParcel = await request.json();
