+++ components/EditorPanel.tsx
@@ -6,19 +6,6 @@ type EditorPanelProps = {
   patchedDocument?: string;
 };
 
-const downloadBlob = (blob: Blob, fileName: string): void => {
-  const link = document.createElement('a');
-  link.href = URL.createObjectURL(blob);
-  link.download = fileName;
-
-  document.body.appendChild(link);
-
-  link.click();
-
-  document.body.removeChild(link);
-  URL.revokeObjectURL(link.href);
-};
-
 export default ({ patchedDocument }: EditorPanelProps) => {
   const router = useRouter();
